+++ src/mnelab/mainwindow.py
@@ -825,6 +825,8 @@ class MainWindow(QMainWindow):
         self.bads = self.model.current["data"].info["bads"]
         events = self.model.current["events"]
         nchan = self.model.current["data"].info["nchan"]
+        nchan = min(nchan, read_settings("max_channels"))
+
         fig = self.model.current["data"].plot(
             events=events,
             n_channels=nchan,